:recycle: Migrate tests into tests directory for volume.rs.
We are migrating the tests into a separate directory so they don't clutter
the source directory. As part of this, we'll be removing/refactoring the
tests that don't work any more. This moves the tests for volume.rs into the
tests directory and disables the running of tests that aren't yet migrated.